+++ src/account/index.ts
@@ -11,9 +11,3 @@ export {
   createAccountStatusBar,
   refreshAccountStatusBar,
 } from "./accountStatusBar";
-export {
-  createUsageStatusBar,
-  refreshUsageStatusBar,
-  setActiveSessionModel,
-  disposeUsageStatusBar,
-} from "./usageStatusBar";